Fix "Aspect Visualization" view
After previous refactoring, the "Aspect Visualization" view in was
broken, throwing errors. This happened due to previous refactoring,
introducing the use of Arrays.asList in order to create lists from
arrays. In most places, this was correct, but in 3 pieces of code, among
them the one causing the aspect visualization to malfunction, I had
overlooked the fact that Arrays.asList returns semi-immutable lists,
in which items can be replaced but not added or removed. Of course,
trying to add/remove items to/from such a list results in errors. This
is now fixed, using 'new ArrayList<>(Arrays.asList(myArray))' instead,
yielding mutable lists.

Signed-off-by: Alexander Kriegisch <Alexander@Kriegisch.name>